Make plugin lookups cheaper
If the plugin is known to come from the registries hierarchy,
look it up there. Otherwise look it up in its own classloader.

The previous logic was more complex and lead to duplicate lookups
if an unknown plugin was requested. This was a performance bottleneck
for Android builds, as the Android plugin makes heavy use of plugin.withId.